refactor: re-organize gdal presets
