+++ scripts/files/file_tiff.py
@@ -6,8 +6,7 @@ from typing import Any
 from urllib.parse import unquote
 
 from scripts.gdal.gdal_helper import GDALExecutionException, gdal_info, run_gdal
-from scripts.gdal.gdal_preset import DEFAULT_NO_DATA_VALUE
-from scripts.gdal.gdal_presets import Preset
+from scripts.gdal.gdal_presets import DEFAULT_NO_DATA_VALUE, Preset
 from scripts.gdal.gdalinfo import GdalInfo
