+++ scripts/gdal/tests/gdal_commands_test.py
@@ -1,7 +1,7 @@
 from pytest_subtests import SubTests
 
 from scripts.gdal.gdal_helper import EpsgNumber
-from scripts.gdal.gdal_preset import get_cutline_command, get_gdal_command
+from scripts.gdal.gdal_commands import get_cutline_command, get_gdal_command
 from scripts.gdal.gdal_presets import Preset
